+++ src/types/github.ts
@@ -1,0 +1,11 @@
+export interface GithubUser {
+  name: string;
+  avatarURL: string;
+  userURL: string;
+}
+
+export interface GithubRawUser {
+  name: string;
+  avatar_url: string;
+  html_url: string;
+}
